Заменил звездочки в форме на рабочий вариант

+++ app/Http/Requests/StoreFeedbackRequest.php
@@ -30,7 +30,7 @@ class StoreFeedbackRequest extends FormRequest
             'surname' => ['string','max:100'],
             'body_text' => ['string'],
             'specialist_id' => ['required','max:200'],
-            'value' => ['integer','min:1','max:5'],
+            'value' => ['numeric','min:1','max:5'],
         ];
     }
 }
